pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "lxml>=6.0.0",
    "pydantic-settings>=2.10.1",
    "pydantic>=2.11.7",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "python-dateutil>=2.9.0.post0",
    "python-multipart>=0.0.20",
    "readability-lxml>=0.8.4.1",
    "slowapi>=0.1.9",
    "trafilatura>=1.6.1",
    "uvicorn>=0.35.0",
    "requests>=2.32.5",
    "openai>=1.100.1",
    "aiohttp>=3.12.15",
]
